Move Change-signature template into a directory and move utility function to own module
